requirements.txt
browser-cookie3==0.6.0
browser-history==0.3.1
cffi==1.15.0
colorama==0.4.4
configparser==4.0.2
cryptography==35.0.0
dominate==2.6.0
hexdump==3.3
importlib-metadata==4.8.1
jeepney==0.7.1
json2html==1.3.0
keyring==23.2.1
LnkParse3==1.2.0
more-itertools==5.0.0
olefile==0.46
pbkdf2==1.3
pyaes==1.6.1
pycparser==2.20
pycryptodome==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-evtx==0.7.4
pywin32==302
pywin32-ctypes==0.2.0
SecretStorage==3.3.1
six==1.16.0
windowsprefetch==4.0.3
WMI==1.5.1
zipp==1.0.0
